backend/requirements.txt
# Core
vanna==0.7.9
anthropic>=0.40.0
chromadb<1.0.0
claude-agent-sdk>=0.1.0

# FastAPI server (TASK 01)
fastapi>=0.110.0
uvicorn>=0.27.0

# Async support
anyio>=4.0.0

# Authentication (TASK 14 - Clean Architecture)
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin to specific version for passlib compatibility
python-multipart
pydantic-settings
email-validator

# Embeddings
langchain-huggingface
sentence-transformers

# Utilities
python-dotenv
flask>=3.0.0
flask-cors
pandas
plotly
requests>=2.31.0

# Testing
pytest
pytest-asyncio
pytest-cov
httpx
